Move date logic from utils to dates.py
To reduce clutter in utils.

Coming in later commits:

- update tests
- update references
